Ensures that missing keys from the config file are set in the internal config
If you would have a `.fasterer.yml` config with only the `speedups` key
pronto would fail because it expects the `exclude_path` key to be there
too. By merging the actual file in the default `nil_config_file` we
ensure that all keys are set internally and default to an empty value
for unset keys.